migrations up working

+++ src/EFCore.MySql/Storage/Internal/MySqlDatabaseCreator.cs
@@ -67,11 +67,11 @@ namespace Microsoft.EntityFrameworkCore.Storage.Internal
         }
 
         protected override bool HasTables()
-            => Dependencies.ExecutionStrategyFactory.Create().Execute(_connection, connection => (int)CreateHasTablesCommand().ExecuteScalar(connection) != 0);
+            => Dependencies.ExecutionStrategyFactory.Create().Execute(_connection, connection => (long)CreateHasTablesCommand().ExecuteScalar(connection) != 0);
 
         protected override Task<bool> HasTablesAsync(CancellationToken cancellationToken = default(CancellationToken))
             => Dependencies.ExecutionStrategyFactory.Create().ExecuteAsync(_connection,
-                async (connection, ct) => (int)await CreateHasTablesCommand().ExecuteScalarAsync(connection, cancellationToken: ct).ConfigureAwait(false) != 0, cancellationToken);
+                async (connection, ct) => (long)await CreateHasTablesCommand().ExecuteScalarAsync(connection, cancellationToken: ct).ConfigureAwait(false) != 0, cancellationToken);
 
         private IRelationalCommand CreateHasTablesCommand()
             => _rawSqlCommandBuilder
